+++ src/com/howfun/android/HF2D/Sprite.java
@@ -14,7 +14,7 @@ public abstract class Sprite {
    protected BitmapDrawable[] mBitmaps;
    protected Pos mPos;
    protected int mType;
-   protected double mSpeed; //Move n pixels per step  
+   protected float mSpeed; //Move n pixels per step  
    
    protected static final int TYPE_ANT = 0;
    protected static final int TYPE_LINE = 1;
